+++ src/by/epamtc/bakulin/service/impl/UserServiceImpl.java
@@ -16,7 +16,8 @@ public class UserServiceImpl implements UserService {
         this.userDAO = userDAO;
     }
 
-    public UserServiceImpl() {}
+    public UserServiceImpl() {
+    }
 
     @Override
     public void setUserDAO(UserDAO userDAO) {
@@ -26,6 +27,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public void addUser(User user) throws ServiceException {
         try {
+            if (user == null) {
+                throw new ServiceException("User value can not be null.");
+            }
             user.setUserId(user.hashCode());
             userDAO.add(user);
         } catch (DAOException e) {
@@ -37,6 +41,9 @@ public class UserServiceImpl implements UserService {
     public User findUserById(Integer id) throws ServiceException {
         User user = null;
         try {
+            if (id == null || id < 0) {
+                throw new ServiceException("Parameter - id, can not be null or less then 0; id = " + id);
+            }
             user = userDAO.findById(id);
         } catch (DAOException e) {
             throw new ServiceException(e);
@@ -48,6 +55,9 @@ public class UserServiceImpl implements UserService {
     public User findUserByName(String userName) throws ServiceException {
         User user = null;
         try {
+            if (userName == null) {
+                throw new ServiceException("Parameter - userName, can not be null");
+            }
             user = userDAO.findByName(userName);
         } catch (DAOException e) {
             throw new ServiceException(e);
@@ -67,8 +77,11 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void updateUser(User user) throws ServiceException{
+    public void updateUser(User user) throws ServiceException {
         try {
+            if (user == null) {
+                throw new ServiceException("User value can not be null.");
+            }
             userDAO.update(user);
         } catch (DAOException e) {
             throw new ServiceException(e);
@@ -78,6 +91,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteUser(Integer id) throws ServiceException {
         try {
+            if (id == null || id < 0) {
+                throw new ServiceException("Parameter - id, can not be null or less then 0; id = " + id);
+            }
             userDAO.delete(id);
         } catch (DAOException e) {
             throw new ServiceException(e);
